Converting amounts to Big Decimal across the board

+++ src/main/java/com/discountify/pojo/DiscountLineItem.java
@@ -1,16 +1,18 @@
 package com.discountify.pojo;
 
+import java.math.BigDecimal;
+
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter @Setter @EqualsAndHashCode
 public class DiscountLineItem {
-	private double amount;
+	private BigDecimal amount;
 	private String description;
 	
 	public DiscountLineItem(){}
-	public DiscountLineItem(String description, double amount){
+	public DiscountLineItem(String description, BigDecimal amount){
 		this.description = description;
 		this.amount = amount;
 	}
